requirements.txt
Pillow==9.1.1
numpy==1.24.3
opencv-python==4.6.0.66
